refactor: merge Checksum.kt into Hash.kt, make internals private
- Extract shared `concatFields` helper (was duplicated between files)
- Make `crc16`, `crc32`, `structToBytes` private (only `computeHash`
  and `onesComplementChecksum` are part of the public API)
- Merge ChecksumTest into HashTest, rewrite internal-function tests
  to go through the public API
