+++ simulator/BUILD.bazel
@@ -88,16 +88,6 @@ kt_jvm_binary(
 # =============================================================================
 
 kt_jvm_test(
-    name = "ChecksumTest",
-    srcs = ["ChecksumTest.kt"],
-    test_class = "fourward.simulator.ChecksumTest",
-    deps = [
-        ":simulator_lib",
-        "@maven//:junit_junit",
-    ],
-)
-
-kt_jvm_test(
     name = "HashTest",
     srcs = ["HashTest.kt"],
     test_class = "fourward.simulator.HashTest",
